requirements.txt
cffi==1.17.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
freetype-py==2.5.1
hsluv==5.0.4
kiwisolver==1.4.7
matplotlib==3.10.0
numpy==1.26.4
numpy-stl==3.2.0
packaging==24.2
pillow==11.0.0
pycparser==2.22
pyFFTW==0.15.0
pyhackrf==0.2.0
PyOpenGL==3.1.7
pyparsing==3.2.0
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
pyqtgraph==0.13.7
pyrtlsdr==0.3.0
python-dateutil==2.9.0.post0
python-utils==3.9.1
scipy==1.14.1
six==1.17.0
sounddevice==0.5.1
typing_extensions==4.12.2
vispy==0.14.3
